+++ packages/blade/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @razorpay/blade
 
+## 11.15.1
+
+### Patch Changes
+
+- 989172ab: fix(blade): tab item button type
+
+## 11.15.0
+
+### Minor Changes
+
+- ada461c8: feat(StepGroup): add StepGroup component
+
+  Documentation: https://blade.razorpay.com/?path=/docs/components-stepgroup--docs
+
 ## 11.14.0
 
 ### Minor Changes